mod: images.js handler. plans for the big opt
